+++ src/main/java/com/keepgoing/keepserver/domain/book/entity/dto/BookRequestDTO.java
@@ -1,4 +1,4 @@
-package com.keepgoing.keepserver.domain.book.repository.dto;
+package com.keepgoing.keepserver.domain.book.entity.dto;
 
 import com.keepgoing.keepserver.domain.book.consts.BookState;
 import lombok.Getter;
@@ -10,5 +10,4 @@ public class BookRequestDTO {
     private String nfcCode;
     private String imageUrl;
     private BookState state;
-
 }
