+++ src/main/java/com/keepgoing/keepserver/domain/book/service/BookService.java
@@ -1,7 +1,7 @@
 package com.keepgoing.keepserver.domain.book.service;
 
 import com.keepgoing.keepserver.domain.book.entity.Book;
-import com.keepgoing.keepserver.domain.book.repository.dto.BookRequestDTO;
+import com.keepgoing.keepserver.domain.book.entity.dto.BookRequestDTO;
 import com.keepgoing.keepserver.global.common.BaseResponse;
 import jakarta.transaction.Transactional;
 import org.springframework.security.core.Authentication;
@@ -14,7 +14,7 @@ public interface BookService {
     BaseResponse selectAllBook();
     BaseResponse deleteBook(String nfcCode, Authentication auth);
     BaseResponse selectMyBook(Authentication auth);
-    public String createNfcCode();
+    String createNfcCode();
     @Transactional
     BaseResponse editBook(String nfcCode, BookRequestDTO bookRequest) throws IOException;
 }
